+++ metrics_utility/automation_controller_billing/package/package_crc.py
@@ -1,26 +1,39 @@
+import os
+import requests
+import json
+import logging
+
 import insights_analytics_collector as base
 
 from awx.main.utils import get_awx_http_client_headers
+from metrics_utility.exceptions import FailedToUploadPayload
 
 from django.conf import settings
-
 
 class PackageCRC(base.Package):
     CERT_PATH = "/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem"
     PAYLOAD_CONTENT_TYPE = "application/vnd.redhat.aap-billing-controller.aap_billing_controller_payload+tgz"
 
+    SHIPPING_AUTH_SERVICE_ACCOUNT = "service-account"
+
     def _tarname_base(self):
         timestamp = self.collector.gather_until
         return f'{settings.SYSTEM_UUID}-{timestamp.strftime("%Y-%m-%d-%H%M%S%z")}'
 
+    def get_sso_url(self):
+        return os.getenv('METRICS_UTILITY_CRC_SSO_URL', 'https://sso.redhat.com/auth/realms/redhat-external/protocol/openid-connect/token')
+
     def get_ingress_url(self):
-        return getattr(settings, 'AUTOMATION_ANALYTICS_URL', None)
+        return os.getenv('METRICS_UTILITY_CRC_INGRESS_URL', 'https://console.redhat.com/api/ingress/v1/upload')
+
+    def get_proxy_url(self):
+        return os.getenv('METRICS_UTILITY_PROXY_URL', None)
 
     def _get_rh_user(self):
-        return getattr(settings, 'REDHAT_USERNAME', None)
+        return os.getenv('METRICS_UTILITY_SERVICE_ACCOUNT_ID', None)
 
     def _get_rh_password(self):
-        return getattr(settings, 'REDHAT_PASSWORD', None)
+        return os.getenv('METRICS_UTILITY_SERVICE_ACCOUNT_SECRET', None)
 
     def _get_http_request_headers(self):
         return get_awx_http_client_headers()
@@ -30,5 +43,91 @@ class PackageCRC(base.Package):
         # for now, uncomment when testin locally in docker
         # return self.SHIPPING_AUTH_IDENTITY
 
-        # TODO: allow to use certificate based auth based on Controller config
-        return self.SHIPPING_AUTH_USERPASS
+        return self.SHIPPING_AUTH_SERVICE_ACCOUNT
+
+    def is_shipping_configured(self):
+        # TODO: move to insights-analytics-collector
+        ret = super()
+        if ret is False:
+            return False
+
+        if self.shipping_auth_mode() == self.SHIPPING_AUTH_SERVICE_ACCOUNT:
+            if not self.get_ingress_url():
+                self.logger.error("METRICS_UTILITY_CRC_INGRESS_URL is not set")
+                return False
+
+            if not self.get_sso_url():
+                self.logger.error("METRICS_UTILITY_CRC_SSO_URL is not set")
+                return False
+
+            if not self._get_rh_user():
+                self.logger.error("METRICS_UTILITY_SERVICE_ACCOUNT_ID is not set")
+                return False
+
+            if not self._get_rh_password():
+                self.logger.error("METRICS_UTILITY_SERVICE_ACCOUNT_SECRET is not set")
+                return False
+        return True
+
+    def _send_data(self, url, files, session):
+        # TODO: move to insights-analytics-collector
+        if self.shipping_auth_mode() == self.SHIPPING_AUTH_SERVICE_ACCOUNT:
+            sso_url = self.get_sso_url()
+            headers = {"Content-Type": "application/x-www-form-urlencoded"}
+
+            data = {"client_id": self._get_rh_user(),
+                    "client_secret": self._get_rh_password(),
+                    "grant_type": "client_credentials"}
+
+            r = requests.post(sso_url,
+                              headers=headers,
+                              data=data,
+                              verify=self.CERT_PATH,
+                              timeout=(31, 31))
+            access_token = json.loads(r.content)['access_token']
+
+            #################################
+            ## Query crc with bearer token
+            headers = session.headers
+            headers['authorization'] = 'Bearer {}'.format(access_token)
+
+            proxies = {}
+            if self.get_proxy_url():
+                proxies = {'https': self.get_proxy_url()}
+
+            response = session.post(
+                url,
+                files=files,
+                verify=self.CERT_PATH,
+                proxies=proxies,
+                headers=headers,
+                timeout=(31, 31),
+            )
+
+        elif self.shipping_auth_mode() == self.SHIPPING_AUTH_USERPASS:
+            response = session.post(
+                url,
+                files=files,
+                verify=self.CERT_PATH,
+                auth=(self._get_rh_user(), self._get_rh_password()),
+                headers=session.headers,
+                timeout=(31, 31),
+            )
+
+        else:
+            response = session.post(
+                url, files=files, headers=session.headers, timeout=(31, 31)
+            )
+
+        # Accept 2XX status_codes
+        if response.status_code >= 300:
+            raise FailedToUploadPayload(
+                "Upload failed with status {}, {}".format(
+                    response.status_code, response.text
+                )
+            )
+
+        return True
+
+
+